requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
